Move more string functions to str.cpp
