+++ src/base/system.h
@@ -1394,15 +1394,6 @@ int str_time(int64_t centisecs, int format, char *buffer, int buffer_size);
 int str_time_float(float secs, int format, char *buffer, int buffer_size);
 
 /**
- * Escapes \ and " characters in a string.
- *
- * @param dst Destination array pointer, gets increased, will point to the terminating null.
- * @param src Source array.
- * @param end End of destination array.
- */
-void str_escape(char **dst, const char *src, const char *end);
-
-/**
  * Utilities for accessing the file system.
  *
  * @defgroup Filesystem
@@ -1651,16 +1642,6 @@ void swap_endian(void *data, unsigned elem_size, unsigned num);
 
 void net_stats(NETSTATS *stats);
 
-int str_toint(const char *str);
-bool str_toint(const char *str, int *out);
-int str_toint_base(const char *str, int base);
-unsigned long str_toulong_base(const char *str, int base);
-int64_t str_toint64_base(const char *str, int base = 10);
-float str_tofloat(const char *str);
-bool str_tofloat(const char *str, float *out);
-
-unsigned str_quickhash(const char *str);
-
 int str_utf8_to_skeleton(const char *str, int *buf, int buf_len);
 
 /**
@@ -1685,21 +1666,6 @@ int str_utf8_comp_confusable(const char *str1, const char *str2);
  * @return Lowercase codepoint, or the original codepoint if there is no lowercase version.
  */
 int str_utf8_tolower_codepoint(int code);
-
-/**
- * Converts the given UTF-8 string to lowercase (locale insensitive).
- *
- * @ingroup Strings
- *
- * @param str String to convert to lowercase.
- * @param output Buffer that will receive the lowercase string.
- * @param size Size of the output buffer.
- *
- * @remark The strings are treated as zero-terminated strings.
- * @remark This function does not work in-place as converting a UTF-8 string to
- *         lowercase may increase its length.
- */
-void str_utf8_tolower(const char *input, char *output, size_t size);
 
 /**
  * Compares two UTF-8 strings case insensitively.
@@ -1761,18 +1727,6 @@ const char *str_utf8_find_nocase(const char *haystack, const char *needle, const
 const char *str_utf8_skip_whitespaces(const char *str);
 
 /**
- * Removes trailing characters that render as spaces by modifying the string in-place.
- *
- * @ingroup Strings
- *
- * @param param Input string.
- *
- * @remark The string is modified in-place.
- * @remark The strings are treated as null-terminated.
- */
-void str_utf8_trim_right(char *param);
-
-/**
  * Moves a cursor forwards in an UTF-8 string.
  *
  * @ingroup Strings
@@ -1786,19 +1740,6 @@ void str_utf8_trim_right(char *param);
  * @remark The strings are treated as null-terminated.
  */
 int str_utf8_forward(const char *str, int cursor);
-
-/**
- * Encode a UTF-8 character.
- *
- * @ingroup Strings
- *
- * @param ptr Pointer to a buffer that should receive the data. Should be able to hold at least 4 bytes.
- *
- * @return Number of bytes put into the buffer.
- *
- * @remark Does not do null-termination of the string.
- */
-int str_utf8_encode(char *ptr, int chr);
 
 /**
  * Checks if a strings contains just valid UTF-8 characters.
@@ -1873,36 +1814,6 @@ size_t str_utf8_offset_bytes_to_chars(const char *str, size_t byte_offset);
  * @remark It's the user's responsibility to make sure the bounds are aligned.
  */
 size_t str_utf8_offset_chars_to_bytes(const char *str, size_t char_offset);
-
-/**
- * Writes the next token after str into buf, returns the rest of the string.
- *
- * @ingroup Strings
- *
- * @param str Pointer to string.
- * @param delim Delimiter for tokenization.
- * @param buffer Buffer to store token in.
- * @param buffer_size Size of the buffer.
- *
- * @return Pointer to rest of the string.
- *
- * @remark The token is always null-terminated.
- */
-const char *str_next_token(const char *str, const char *delim, char *buffer, int buffer_size);
-
-/**
- * Checks if needle is in list delimited by delim.
- *
- * @param list List.
- * @param delim List delimiter.
- * @param needle Item that is being looked for.
- *
- * @return `1` - Item is in list.
- * @return `0` - Item isn't in list.
- *
-* @remark The strings are treated as null-terminated strings.
- */
-int str_in_list(const char *list, const char *delim, const char *needle);
 
 /**
  * Packs 4 big endian bytes into an unsigned.
